+++ R/geom-grob.R
@@ -1,30 +1,37 @@
 #' Inset graphical objects
 #'
 #' \code{geom_grob} and \code{geom_grob_npc} add Grobs as insets to the ggplot
-#' using syntax similar to that of \code{\link[ggplot2]{geom_text}} and
-#' \code{\link{geom_text_s}}. In most respects they behave as any other ggplot
-#' geometry: ther add a layer containing one or more grobs and faceting works as
-#' usual.
+#' using syntax similar to that of \code{\link[ggplot2]{geom_text}},
+#' \code{\link{geom_text_s}} and \code{\link{geom_text_npc}}.
+#' In most respects they behave as any other ggplot geometry: they add a layer
+#' containing one or more grobs and grouping and faceting works as usual. The
+#' most common use of \code{geom_grob} is to add data labels that are graphical
+#' objects rather than text. \code{\link{geom_grob_npc}} is used to add grobs
+#' as annotations to plots, but contrary to layer function \code{annotate()},
+#' \code{\link{geom_grob_npc}} is data driven and respects grouping and facets,
+#' thus plot insets can differ among panels.
 #'
-#' @details You can modify the size of inset grobs with the \code{vp.width} and
+#' @details You can modify the size of insets with the \code{vp.width} and
 #'   \code{vp.height} aesthetics. These can take a number between 0 (smallest
 #'   possible inset) and 1 (whole plotting area width or height). The default
 #'   value for for both of these aesthetics is 1/5. Thus, in contrast to
-#'   \code{\link[ggplot2]{geom_text}} and \code{\link{geom_text_s}} the size of
-#'   the insets remains the same relative to the size of the plotting area
-#'   irrespective of how the plot is rendered. The aspect ratio of insets is
-#'   preserved and size is adjusted until the whole inset fits within the
-#'   viewport.
+#'   \code{\link[ggplot2]{geom_text}}, \code{\link[ggplot2]{geom_label}},
+#'   \code{\link{geom_text_s}} and \code{\link{geom_label_s}} the size of the
+#'   insets remains the same relative to the size of the plotting area
+#'   irrespective of the size the plot is rendered at. The aspect ratio of
+#'   insets is preserved and size is adjusted until the whole inset fits within
+#'   the viewport.
 #'
-#'   By default this geom uses \code{\link{position_nudge_center}} which is
-#'   backwards compatible with \code{\link[ggplot2]{position_nudge}} but
-#'   provides additional control on the direction of the nudging. In contrast to
-#'   \code{\link[ggplot2]{position_nudge}}, \code{\link{position_nudge_center}}
-#'   and all other position functions defined in packages 'ggpp' and 'ggrepel'
-#'   keep the original coordinates thus allowing the plotting of connecting
-#'   segments and arrows.
+#'   By default \code{geom_grob} uses \code{\link{position_nudge_center}} and
+#'   justification \code{"position"}, while \code{geom_grob_npc} uses
+#'   \code{\link[ggplot2]{position_nudge}} and justification \code{"inward"}. In
+#'   contrast to \code{\link[ggplot2]{position_nudge}},
+#'   \code{\link{position_nudge_center}} and all other position functions
+#'   defined in packages 'ggpp' keep the original coordinates thus allowing the
+#'   plotting of connecting segments and arrows.
 #'
-#'   This geom works only with tibbles as \code{data}, as its expects a list of
+#'   This \code{geom_grob} and \code{geom_grob_npc} require the use tibbles as
+#'   argument for \code{data}, as the grobs should be stored as a list of
 #'   graphics objects ("grob") to be mapped to the \code{label} aesthetic.
 #'
 #'   The \code{x} and \code{y} aesthetics determine the position of the whole
@@ -33,36 +40,35 @@
 #'   \emph{y} coordinates in the data, and \code{angle} is used to rotate the
 #'   grob as a whole.
 #'
-#'   In the case of \code{geom_table_npc()}, \code{npcx} and \code{npcy}
-#'   aesthetics determine the position of the inset table. Justification is as
-#'   described above for \code{geom_table()}.
+#' @section Plot boundaries and clipping: The "width" and "height" of an inset
+#'   as for a text element are 0, so stacking and dodging inset plots will not
+#'   work by default, and axis limits are not automatically expanded to include
+#'   all inset plots. Obviously, insets do have height and width, but they are
+#'   physical units, not data units. The amount of space they occupy on the main
+#'   plot is not constant in data units of the base plot: when you modify scale
+#'   limits, inset plots stay the same size relative to the physical size of the
+#'   base plot.
 #'
-#'   Use \code{\link{annotate}} as redefined in 'ggpp' when adding inset grobs
+#' @inheritSection geom_text_s Alignment
+#'
+#' @inheritSection geom_text_s Position functions
+#'
+#' @note The insets are stored nested within the main ggplot object and
+#'   contain their own copy of the data, and are rendered as grid grobs as normal
+#'   ggplots at the time the main ggplot is rendered. They can have different
+#'   themes.
+#'
+#'   Use \code{\link{annotate}} as redefined in 'ggpp' when adding insets
 #'   as annotations (automatically available unless 'ggpp' is not attached).
-#'   \code{\link[ggplot2]{annotate}} cannot be used with \code{geom = "grob"}.
+#'   \code{\link[ggplot2]{annotate}} cannot be used with the \code{npcx} and
+#'   \code{npcy} pseudo-aesthetics.
 #'
-#' @section Alignment: You can modify the alignment of the grob with the `vjust`
-#'   and `hjust` aesthetics. These can either be a number between 0
-#'   (right/bottom) and 1 (top/left) or a character (\code{"left"},
-#'   \code{"middle"}, \code{"right"}, \code{"bottom"}, \code{"center"},
-#'   \code{"top"}). In addition, you can use special alignments for
-#'   justification including \code{"inward"} and \code{"outward"}. Inward always
-#'   aligns text towards the center of the plotting area, and outward aligns it
-#'   away from the center of the plotting area. If tagged with \code{_mean} or
-#'   \code{_median} (e.g., \code{"outward_mean"}) the mean or median of the data
-#'   in the panel along the corresponding axis is used as center. If the
-#'   characters following the underscore represent a number (e.g.,
-#'   \code{"outward_10.5"}) the reference point will be this value in data
-#'   units.
-#'
-#' @seealso \code{\link{geom_plot}}, \code{\link{geom_table}},
-#'   \code{\link{annotate}}, \code{\link{position_nudge_keep}},
-#'   \code{\link{position_nudge_to}}, \code{\link{position_jitternudge}},
-#'   \code{\link{position_dodgenudge}} and \code{\link{position_stacknudge}}.
+#' @seealso \code{\link[grid]{grid-package}}, \code{\link[ggplot2]{geom_text}},
+#'   and other documentation of package 'ggplot2'.
 #'
 #' @param mapping The aesthetic mapping, usually constructed with
-#'   \code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs to be
-#'   set at the layer level if you are overriding the plot defaults.
+#'   \code{\link[ggplot2]{aes}}. Only needs to be set at the layer level if you
+#'   are overriding the plot defaults.
 #' @param data A layer specific dataset - only needed if you want to override
 #'   the plot defaults.
 #' @param stat The statistical transformation to use on the data for this layer,
@@ -84,8 +90,22 @@
 #' @param nudge_x,nudge_y Horizontal and vertical adjustments to nudge the
 #'   starting position of each text label. The units for \code{nudge_x} and
 #'   \code{nudge_y} are the same as for the data units on the x-axis and y-axis.
+#' @param default.colour A colour definition to use for elements not targeted by
+#'   the colour aesthetic.
+#' @param colour.target A vector of character strings; \code{"all"},
+#'   \code{"text"}, \code{"box"} and \code{"segment"}.
+#' @param default.alpha numeric in [0..1] A transparency value to use for
+#'   elements not targeted by the alpha aesthetic.
+#' @param alpha.target A vector of character strings; \code{"all"},
+#'   \code{"text"}, \code{"segment"}, \code{"box"}, \code{"box.line"}, and
+#'   \code{"box.fill"}.
 #' @param add.segments logical Display connecting segments or arrows between
 #'   original positions and displaced ones if both are available.
+#' @param box.padding,point.padding numeric By how much each end of the segments
+#'   should shortened in mm.
+#' @param segment.linewidth numeric Width of the segments or arrows in mm.
+#' @param min.segment.length numeric Segments shorter that the minimum length
+#'   are not rendered, in mm.
 #' @param arrow specification for arrow heads, as created by
 #'   \code{\link[grid]{arrow}}
 #'
@@ -95,7 +115,7 @@
 #'   been discussed as an issue at
 #'   \url{https://github.com/tidyverse/ggplot2/issues/1399}.
 #'
-#' @family geometries adding layers with insets
+#' @family geometries adding layers with insets.
 #'
 #' @export
 #'
@@ -117,7 +137,7 @@
 #'   geom_grob(data = df,
 #'             aes(x, y, label = grob),
 #'             nudge_x = 0.5,
-#'             segment.colour = "red")
+#'             colour = "red")
 #'
 #' # with nudging plotting of segments can be disabled
 #' ggplot(data = mtcars,
@@ -135,7 +155,15 @@ geom_grob <- function(mapping = NULL,
                       ...,
                       nudge_x = 0,
                       nudge_y = 0,
+                      default.colour = "black",
+                      colour.target = "segment",
+                      default.alpha = 1,
+                      alpha.target = "segment",
                       add.segments = TRUE,
+                      box.padding = 0.25,
+                      point.padding = 1e-06,
+                      segment.linewidth = 0.5,
+                      min.segment.length = 0,
                       arrow = NULL,
                       na.rm = FALSE,
                       show.legend = FALSE,
@@ -145,11 +173,9 @@ geom_grob <- function(mapping = NULL,
     if (!missing(position) && position != "identity") {
       rlang::abort("You must specify either `position` or `nudge_x`/`nudge_y`.")
     }
-    # We do not keep the original positions if they will not be used
+    # original position needed for "position" justification
     position <-
-      position_nudge_center(nudge_x, nudge_y,
-                            kept.origin = ifelse(add.segments,
-                                                 "original", "none"))
+      position_nudge_center(nudge_x, nudge_y, kept.origin = "original")
   }
 
   ggplot2::layer(
@@ -161,7 +187,15 @@ geom_grob <- function(mapping = NULL,
     show.legend = show.legend,
     inherit.aes = inherit.aes,
     params = list(
+      default.colour = default.colour,
+      colour.target = colour.target,
+      default.alpha = default.alpha,
+      alpha.target = alpha.target,
       add.segments = add.segments,
+      box.padding = box.padding,
+      point.padding = point.padding,
+      segment.linewidth = segment.linewidth,
+      min.segment.length = min.segment.length,
       arrow = arrow,
       na.rm = na.rm,
       ...
@@ -178,8 +212,16 @@ grob_draw_panel_fun <-
   function(data,
            panel_params,
            coord,
+           default.colour = "black",
+           colour.target = "all",
+           default.alpha = 1,
+           alpha.target = "all",
            na.rm = FALSE,
            add.segments = TRUE,
+           box.padding = 0.25,
+           point.padding = 1e-06,
+           segment.linewidth = 1,
+           min.segment.length = 0,
            arrow = NULL) {
 
     if (nrow(data) == 0) {
@@ -218,12 +260,25 @@ grob_draw_panel_fun <-
                        just = data$hjust,
                        a = "x", b = "y")
     }
+    if (add.segments) {
+      segments.data <-
+        shrink_segments(data,
+                        point.padding = point.padding,
+                        box.padding = box.padding,
+                        min.segment.length = min.segment.length)
+    }
 
     # loop needed as gpar is not vectorized
     all.grobs <- grid::gList()
     user.grobs <- data[["label"]]
     for (row.idx in 1:nrow(data)) {
       row <- data[row.idx, , drop = FALSE]
+      grob.alpha <-
+        ifelse(any(alpha.target %in% c("all", "grob")),
+               row$alpha, default.alpha)
+      segment.alpha <-
+        ifelse(any(alpha.target %in% c("all", "segment")),
+               row$alpha, default.alpha)
       user.grob <- user.grobs[[row.idx]]
 
       user.grob$vp <-
@@ -240,16 +295,25 @@ grob_draw_panel_fun <-
       user.grob$name <- paste("inset.grob", row.idx, sep = ".")
 
       if (add.segments) {
+        segment.row <- segments.data[row.idx, , drop = FALSE]
+        if (segment.row$too.short) {
+          segment.grob <- grid::nullGrob()
+        } else {
           segment.grob <-
-          grid::segmentsGrob(x0 = row$x,
-                             y0 = row$y,
-                             x1 = row$x_orig,
-                             y1 = row$y_orig,
-                             arrow = arrow,
-                             gp = grid::gpar(col = ggplot2::alpha(row$segment.colour,
-                                                                  row$segment.alpha)),
-                             name = paste("inset.grob.segment", row.idx, sep = "."))
-          all.grobs <- grid::gList(all.grobs, segment.grob, user.grob)
+            grid::segmentsGrob(x0 = segment.row$x,
+                               y0 = segment.row$y,
+                               x1 = segment.row$x_orig,
+                               y1 = segment.row$y_orig,
+                               arrow = arrow,
+                               gp = grid::gpar(
+                                 col = if (segment.linewidth == 0) NA else # lwd = 0 is invalid in 'grid'
+                                   ifelse(any(colour.target %in% c("all", "segment")),
+                                          ggplot2::alpha(row$colour, segment.alpha),
+                                          ggplot2::alpha(default.colour, segment.alpha)),
+                                 lwd = (if (segment.linewidth == 0) 1 else segment.linewidth) * .stroke),
+                               name = paste("grob.s.segment", row$group, row.idx, sep = "."))
+        }
+        all.grobs <- grid::gList(all.grobs, segment.grob, user.grob)
       } else {
         all.grobs <- grid::gList(all.grobs, user.grob)
       }
@@ -268,13 +332,14 @@ GeomGrob <-
                    required_aes = c("x", "y", "label"),
 
                    default_aes = ggplot2::aes(
-                     colour = "black", angle = 0, hjust = 0.5,
-                     vjust = 0.5, alpha = NA, family = "", fontface = 1,
-                     vp.width = 1/5, vp.height = 1/5,
-                     segment.linetype = 1,
-                     segment.colour = "grey33",
-                     segment.size = 0.5,
-                     segment.alpha = 1
+                     colour = "black",
+                     angle = 0,
+                     hjust = "position",
+                     vjust = "position",
+                     alpha = NA,
+                     family = "",
+                     fontface = 1,
+                     vp.width = 1/5, vp.height = 1/5
                    ),
 
                    draw_panel = grob_draw_panel_fun,
@@ -316,7 +381,9 @@ geom_grob_npc <- function(mapping = NULL,
 #' @usage NULL
 #'
 grobnpc_draw_panel_fun <-
-  function(data, panel_params, coord,
+  function(data,
+           panel_params,
+           coord,
            na.rm = FALSE) {
 
     if (nrow(data) == 0) {