+++ R/geom-table.R
@@ -3,8 +3,13 @@
 #' \code{geom_table} and \code{geom_table_npc} add data frames as table insets
 #' to the base ggplot, using syntax similar to that of
 #' \code{\link[ggplot2]{geom_text}} and \code{\link{geom_text_s}}. In most
-#' respects they behave as any other ggplot geometry: a layer con contain
-#' multiple tables and faceting works as usual.
+#' respects they behave as any other ggplot geometry: they add a layer
+#' containing one or more grobs and grouping and faceting works as usual. The
+#' most common use of \code{geom_table} is to add data labels that are whole
+#' tables rather than text. \code{\link{geom_table_npc}} is used to add tables
+#' as annotations to plots, but contrary to layer function \code{annotate},
+#' \code{\link{geom_table_npc}} is data driven and respects grouping and facets,
+#' thus plot insets can differ among panels.
 #'
 #' @details You can modify the size of inset tables with the \code{vp.width} and
 #'   \code{vp.height} aesthetics. These can take a number between 0 (smallest
@@ -26,7 +31,7 @@
 #'
 #'   This geom works only with tibbles as \code{data}, as its expects a list of
 #'   data frames (or tibbles) to be mapped to the \code{label} aesthetic.
-#'   A table is built with function \code{gridExtra::gtable()} for each
+#'   A table is built with function \code{gridExtra::gtable} for each
 #'   data frame in the list, and formatted according to a table theme or
 #'   \code{ttheme}. The character strings in the data frame can be parsed into
 #'   R expressions so the inset tables can include maths.
@@ -51,34 +56,30 @@
 #'   its \emph{horizontal} and \emph{vertical} axes (rows and columns in the
 #'   data frame), and \code{angle} is used to rotate the inset table as a whole.
 #'
-#'   In the case of \code{geom_table_npc()}, \code{npcx} and \code{npcy}
+#'   In the case of \code{geom_table_npc}, \code{npcx} and \code{npcy}
 #'   aesthetics determine the position of the inset table. Justification as
 #'   described above for .
 #'
-#'   Use \code{\link{annotate}} as redefined in 'ggpp' when adding inset plots
-#'   as annotations (automatically available unless 'ggpp' is not attached).
-#'   \code{\link[ggplot2]{annotate}} cannot be used with \code{geom = "table"}.
+#' @inheritSection geom_text_s Alignment
 #'
-#' @section Alignment: You can modify the alignment of the whole table with the `vjust` and
-#'   `hjust` aesthetics. These can either be a number between 0 (right/bottom)
-#'   and 1 (top/left) or a character (\code{"left"}, \code{"middle"},
-#'   \code{"right"}, \code{"bottom"}, \code{"center"}, \code{"top"}). In
-#'   addition, you can use special alignments for justification including
-#'   \code{"inward"} and \code{"outward"}. Inward always aligns text towards
-#'   the center of the plotting area, and outward aligns it away from the center
-#'   of the plotting area. If tagged with \code{_mean} or \code{_median} (e.g.,
-#'   \code{"outward_mean"}) the mean or median of the data in the panel along
-#'   the corresponding axis is used as center. If the characters following the
-#'   underscore represent a number (e.g., \code{"outward_10.5"}) the reference
-#'   point will be this value in data units.
+#' @inheritSection geom_text_s Position functions
+#'
+#' @inheritSection geom_grob Plot boundaries and clipping
+#'
+#' @note Complex tables with annotations or different colouring of rows or cells
+#'   can be constructed with functions in package 'gridExtra' or in any other
+#'   way as long as they can be saved as grid graphical objects and then added
+#'   to a ggplot as a new layer with \code{\link{geom_grob}}.
+#'
+#' @inherit geom_grob note return seealso references
 #'
 #' @seealso Formatting of tables \code{stat_fmt_table},
 #'   \code{\link{ttheme_gtdefault}}, \code{\link{ttheme_set}},
 #'   \code{\link[gridExtra]{tableGrob}}.
 #'
 #' @param mapping The aesthetic mapping, usually constructed with
-#'   \code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs
-#'   to be set at the layer level if you are overriding the plot defaults.
+#'   \code{\link[ggplot2]{aes}}. Only needs to be set at the layer level if you
+#'   are overriding the plot defaults.
 #' @param data A layer specific data set - only needed if you want to override
 #'   the plot defaults.
 #' @param stat The statistical transformation to use on the data for this layer,
@@ -108,15 +109,24 @@
 #' @param nudge_x,nudge_y Horizontal and vertical adjustments to nudge the
 #'   starting position of each text label. The units for \code{nudge_x} and
 #'   \code{nudge_y} are the same as for the data units on the x-axis and y-axis.
+#' @param default.colour A colour definition to use for elements not targeted by
+#'   the colour aesthetic.
+#' @param colour.target A vector of character strings; \code{"all"},
+#'   \code{"text"}, \code{"box"} and \code{"segment"}.
+#' @param default.alpha numeric in [0..1] A transparency value to use for
+#'   elements not targeted by the alpha aesthetic.
+#' @param alpha.target A vector of character strings; \code{"all"},
+#'   \code{"text"}, \code{"segment"}, \code{"box"}, \code{"box.line"}, and
+#'   \code{"box.fill"}.
 #' @param add.segments logical Display connecting segments or arrows between
 #'   original positions and displaced ones if both are available.
+#' @param box.padding,point.padding numeric By how much each end of the segments
+#'   should shortened in mm.
+#' @param segment.linewidth numeric Width of the segments or arrows in mm.
+#' @param min.segment.length numeric Segments shorter that the minimum length
+#'   are not rendered, in mm.
 #' @param arrow specification for arrow heads, as created by
 #'   \code{\link[grid]{arrow}}
-#'
-#' @note Complex tables with annotations or different colouring of rows or cells
-#'   can be constructed with functions in package 'gridExtra' or in any other
-#'   way as long as they can be saved as grid graphical objects and then added
-#'   to a ggplot as a new layer with \code{\link{geom_grob}}.
 #'
 #' @references This geometry is inspired on answers to two questions in
 #'   Stackoverflow. In contrast to these earlier examples, the current geom
@@ -126,8 +136,6 @@
 #'   \url{https://stackoverflow.com/questions/25554548/adding-sub-tables-on-each-panel-of-a-facet-ggplot-in-r?}
 #'
 #' @family geometries adding layers with insets
-#'
-#' @return A plot layer instance.
 #'
 #' @export
 #'
@@ -218,7 +226,15 @@ geom_table <- function(mapping = NULL, data = NULL,
                        ...,
                        nudge_x = 0,
                        nudge_y = 0,
+                       default.colour = "black",
+                       colour.target = "segment",
+                       default.alpha = 1,
+                       alpha.target = "segment",
                        add.segments = TRUE,
+                       box.padding = 0.25,
+                       point.padding = 1e-06,
+                       segment.linewidth = 0.5,
+                       min.segment.length = 0,
                        arrow = NULL,
                        table.theme = NULL,
                        table.rownames = FALSE,
@@ -258,7 +274,15 @@ geom_table <- function(mapping = NULL, data = NULL,
     show.legend = show.legend,
     inherit.aes = inherit.aes,
     params = list(
+      default.colour = default.colour,
+      colour.target = colour.target,
+      default.alpha = default.alpha,
+      alpha.target = alpha.target,
       add.segments = add.segments,
+      box.padding = box.padding,
+      point.padding = point.padding,
+      segment.linewidth = segment.linewidth,
+      min.segment.length = min.segment.length,
       arrow = arrow,
       table.theme = table.theme,
       table.rownames = table.rownames,
@@ -283,12 +307,20 @@ gtb_draw_panel_fun <-
            panel_params,
            coord,
            add.segments = TRUE,
+           box.padding = 0.25,
+           point.padding = 1e-06,
+           segment.linewidth = 1,
+           min.segment.length = 0,
            arrow = NULL,
            table.theme = NULL,
            table.rownames = FALSE,
            table.colnames = TRUE,
            table.hjust = 0.5,
            parse = FALSE,
+           default.colour = "black",
+           colour.target = "all",
+           default.alpha = 1,
+           alpha.target = "all",
            na.rm = FALSE) {
 
     if (nrow(data) == 0) {
@@ -328,6 +360,13 @@ gtb_draw_panel_fun <-
                        just = data$hjust,
                        a = "x", b = "y")
     }
+    if (add.segments) {
+      segments.data <-
+        shrink_segments(data,
+                        point.padding = point.padding,
+                        box.padding = box.padding,
+                        min.segment.length = min.segment.length)
+    }
 
     # replace NULL with default
     if (is.null(table.theme)) {
@@ -340,6 +379,12 @@ gtb_draw_panel_fun <-
 
     for (row.idx in 1:nrow(data)) {
       row <- data[row.idx, , drop = FALSE]
+      table.alpha <-
+        ifelse(any(alpha.target %in% c("all", "table")),
+               row$alpha, default.alpha)
+      segment.alpha <-
+        ifelse(any(alpha.target %in% c("all", "segment")),
+               row$alpha, default.alpha)
 
       # Build the table
       if (is.function(table.theme)) {
@@ -404,22 +449,30 @@ gtb_draw_panel_fun <-
       user.grob$name <- paste("inset.table", row.idx, sep = ".")
 
       if (add.segments) {
-        segment.grob <-
-          grid::segmentsGrob(x0 = row$x,
-                             y0 = row$y,
-                             x1 = row$x_orig,
-                             y1 = row$y_orig,
-                             arrow = arrow,
-                             gp = grid::gpar(col = ggplot2::alpha(row$segment.colour,
-                                                                  row$segment.alpha)),
-                             name = paste("inset.table.segment", row.idx, sep = "."))
+        segment.row <- segments.data[row.idx, , drop = FALSE]
+        if (segment.row$too.short) {
+          segment.grob <- grid::nullGrob()
+        } else {
+          segment.grob <-
+            grid::segmentsGrob(x0 = segment.row$x,
+                               y0 = segment.row$y,
+                               x1 = segment.row$x_orig,
+                               y1 = segment.row$y_orig,
+                               arrow = arrow,
+                               gp = grid::gpar(
+                                 col = if (segment.linewidth == 0) NA else # lwd = 0 is invalid in 'grid'
+                                   ifelse(any(colour.target %in% c("all", "segment")),
+                                          ggplot2::alpha(row$colour, segment.alpha),
+                                          ggplot2::alpha(default.colour, segment.alpha)),
+                                 lwd = (if (segment.linewidth == 0) 1 else segment.linewidth) * .stroke),
+                               name = paste("table.s.segment", row$group, row.idx, sep = "."))
+        }
         all.grobs <- grid::gList(all.grobs, segment.grob, user.grob)
       } else {
         all.grobs <- grid::gList(all.grobs, user.grob)
       }
     }
-
-#    grid::grobTree(children = all.grobs, name = "geom.table.panel")
+  #    grid::grobTree(children = all.grobs, name = "geom.table.panel")
     grid::grobTree(children = all.grobs)
 
   }
@@ -442,11 +495,7 @@ GeomTable <-
             alpha = 1,
             family = "",
             fontface = 1,
-            lineheight = 1.2,
-            segment.linetype = 1,
-            segment.colour = "grey33",
-            segment.size = 0.5,
-            segment.alpha = 1
+            lineheight = 1.2
           ),
 
           draw_panel = gtb_draw_panel_fun,
@@ -619,9 +668,15 @@ GeomTableNpc <-
           required_aes = c("npcx", "npcy", "label"),
 
           default_aes = ggplot2::aes(
-            colour = NA, fill = NA,
-            size = 3.2, angle = 0, hjust = "inward",
-            vjust = "inward", alpha = 1, family = "", fontface = 1,
+            colour = NA,
+            fill = NA,
+            size = 3.2,
+            angle = 0,
+            hjust = "inward",
+            vjust = "inward",
+            alpha = 1,
+            family = "",
+            fontface = 1,
             lineheight = 1.2
           ),